Create useAuth hook for logging in
This may not be the most idiomatic React or idiomatic usage of hooks,
but the conversion seems to be working. Some things to look at:

To auth on mount, an empty array is passed into useEffect in useAuth.

The Login component now takes children, which feels more consistent
conceptually, but maybe can be simplified with more liberal use of hooks.

useAuth returns [{uid, owner}, {authenticate, logout}]. It might
be best to create another higher level hook that returns something
like { isOwner, isLoggedIn }.